Create Islands.
Islands are navigation meshes that can be swapped in and out. An Archipelago can also be made of several Islands, so multiple nav meshes can be used at once.

In this commit, swapping out a nav mesh will not behave correctly (paths will be nonsensical), and there is no way to walk between Islands, but the next few commits will fix that.