Create mod geometry and create edge_intersection.
This is meant to be a mod for all the heavy geometry computations that don't have a good home and need thorough testing. I decided to keep this separate from util since util is likely to grow in size and geometry is a big enough portion that keeping it separate seems nice.