Copy `AgentState` from `landmass` and remove `AgentCurrentState`.
`AgentCurrentState` existed because I didn't want to copy the `landmass` enum but I needed a component to hold it. I have decided copying this is fine. It means users can change the state now (which does nothing), but now there's no confusion about what the difference between `AgentCurrentState` and `AgentState` is (although at the cost of needing to keep the structs in sync).