Use nav mesh vertices for modified nodes where close enough.
Previously, a modified node had to have all vertices be arbitrary. Now only the truly new ones are added as Vec2's and the rest are just indices in the original nav mesh. This may seem like an optimization but it actually makes it simpler to implement avoidance since we can rely on basic index math instead of comparing Vec2's.