Fix pathfinding having a wrong heuristic and looking up the wrong island.
Originally pathfinding was written for a single island (back when islands didn't exist!). It seems while I was updating it to support islands I still assumed everything was on a single island (e.g., the end node was on the same island as the state node).

I also assumed that node connections were always on the same island as the start node.