Remove the obstacle_avoidance_margin.
Turns out I can just set this to zero! I am silly. This fixes an issue where characters rounding a corner would slow down **a lot**. Now they just round the corner at full speed as expected. Note that this works because we assume that the nav mesh is the "valid" region, so the avoidance obstacles are really an "outer bound" rather than acting like walls.