Move the connectivity data into the polygons and make it per-edge.
Previously all the connectivity data was in a separate structure **and** the connectivity was not indexed - there was no way to tell whether a particular edge is connected. To solve this, we can just a vector of Options instead, and store this in each polygon. Now we can check if a particular edge is connected and iterate through the connections just the same. We could use something like a HashMap, but it's likely more performant to just use a Vec instead.