Exit early from pathfinding if the regions of nodes are not connected.
Previously if two nodes had **no** path between them, the search would look through the **entire** connected graph before giving up. Now, if there is no path, we immediately exit!

This still leaves open the door that if there is a path, but one that takes a very long detour, the search could explore nearly all of the graph before finding any path at all. Another problem is that if there is no path, agents will just sit and do nothing. Instead, we may want to find a path getting "close" to the target, and follow that.