Fix crashing on empty nav meshes.
We could consider empty nav meshes invalid (and I may still do that), but I still don't like the fact that we're crashing for them. This is a simple extra filter, so no big deal. We can still check for empty nav meshes in `NavigationMesh::validate` later if we decide we don't want to support that.