Make Archipelago generic on the coordinate system.
Now pretty much everything in landmass is generic on the coordinate
system that needs to be. There's still a question of whether to make
ValidNavigationMesh generic on the coordinate system. It's not
technically necessary, but it would add more type safety (you can't
accidentally use a 2D navigation mesh in a 3D archipelago).