Avoid updating the nav data entirely if the islands haven't been changed.
Now we know whether the nav data is in a "stable" state (unchanged since the last update). For one, we can entirely avoid updating the nav data if nothing is changed. No need to poll for changes! This also unlocks the door for future changes that require that the nav mesh be in a valid state (e.g., debugging).