Make `find_path` to find a path along the archipelago.
Now users can find paths adhoc! This is useful if you only want your AI character to "know" if something is nearby if it's within some (walkable) range. For example, Half-Life used this on monsters to "smell" the player, by checking if the player is some walkable distance away. This prevented smelling through walls!