Make the pathfinding heuristic use the cheapest node cost.
The heuristic needs to be "admissible" for the path to be optimal. This means a lot of things, but the key thing we're missing with node types is that the heuristic must be **less than the actual cost of the path**. If all node types were the cheapest cost, the cost can be lower than the previous heuristic, making it inadmissable.