Make `Archipelago::new` take an `AgentOptions`.
This makes it an "obstacle" that prevents a mistake where the defaults don't make any sense due to the size. For example, in 2D, users (like me) may use pixels, and so a character could be ~100px in size. This results in strange behavior since then the neighbourhood is 5px according to the defaults.