Make the schedule configurable and change the default to RunFixedMainLoop (#115)
Previously, we ran the navigation logic in Update - this occurs after FixedUpdate which is where physics movement (usually) occurs. Ideally we'd like to integrate as many changes and assign the most up-to-date desired movement before FixedUpdate. Putting it in RunFixedMainLoop allows us to do this.

In addition, this allows users to configure the schedule if that is not desirable. This does mean they will have an extraneous system set dependency in another schedule, but practically this won't matter.