Fix a flake in `switching_nav_mesh_to_fewer_vertices_does_not_result_in_panic` test.
This test was flaky seemingly because of some floating point inaccuracy. I don't really understand how that could be non-deterministic, but whatever. The solution is to just move the agent further from the corner that is causing the agent to slow down.

An alternative solution could be to adjust the agent's velocity, but I'd rather not do that.