Replace node types with just directly using the type index in `landmass`.
Node types are an ok idea if using `landmass` directly, but for example `bevy_landmass` makes it an incredible pain to use node types. Since we can only create node types from a spawned `Archipelago`, that meant we needed to create the node types in a system and then somehow pass that to an asset loader in order to load the nav meshes with the type index directly in them.

In addition to those ergonomics problems, most existing navigation systems seem to get by totally fine with just an integer for areas, so we can too!